move int module out from internal to Marshal.
even though this is an internal module, everything else now should be
considered internal.
